refactored eval modules
