+++ src/prml_vslam/align/__init__.py
@@ -1,0 +1,13 @@
+"""Alignment algorithms: gravity detection, Sim(3) trajectory, ICP point cloud."""
+
+from .gravity import is_gravity_aligned_target
+from .icp import CloudAlignmentService
+from .trajectory_sim3 import align_estimate_sim3, sim3_up_axis_tilt_deg, trajectory_supports_sim3
+
+__all__ = [
+    "CloudAlignmentService",
+    "align_estimate_sim3",
+    "is_gravity_aligned_target",
+    "sim3_up_axis_tilt_deg",
+    "trajectory_supports_sim3",
+]
