+++ src/prml_vslam/eval/services.py
@@ -4,30 +4,32 @@ This module implements the explicit evaluation work described by
 :mod:`prml_vslam.eval.contracts` and :mod:`prml_vslam.eval.protocols`. App and
 post-run aggregation discovery lives in :mod:`prml_vslam.eval.query`; this
 module owns eval-stage computation and persistence.
+
+Alignment algorithms (Sim3, ICP, gravity detection) live in
+:mod:`prml_vslam.align`.
 """
 
 from __future__ import annotations
 
-import copy
 import csv
 import json
 import math
 from dataclasses import dataclass
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import numpy as np
-import open3d as o3d
 from evo.core import metrics, sync
-from evo.core.trajectory import PoseTrajectory3D
 from evo.tools import file_interface
 
+from prml_vslam.align.gravity import _RDF_DOWN_AXIS
+from prml_vslam.align.trajectory_sim3 import align_estimate_sim3, sim3_up_axis_tilt_deg, trajectory_supports_sim3
 from prml_vslam.eval.alignment_contracts import (
     TrajectoryAlignmentArtifact,
     TrajectoryAlignmentCloudUseStatus,
     TrajectoryAlignmentMode,
 )
-from prml_vslam.eval.contracts import CloudAlignmentArtifact, CloudAlignmentSelection, MetricStats
+from prml_vslam.eval.contracts import MetricStats
 from prml_vslam.eval.trajectory_contracts import (
     DiscoveredRun,
     SelectionSnapshot,
@@ -40,16 +42,13 @@ from prml_vslam.interfaces.slam import SlamArtifacts
 from prml_vslam.sources.contracts import PreparedBenchmarkInputs, ReferenceTrajectoryRef, SequenceManifest
 from prml_vslam.sources.datasets.contracts import DatasetId
 from prml_vslam.utils.geometry import (
-    apply_similarity_to_trajectory,
     load_point_cloud_ply_with_colors,
     load_tum_trajectory,
     write_point_cloud_ply,
-    yaw_similarity_align,
 )
 from prml_vslam.utils.path_config import PathConfig
 
 __all__ = [
-    "CloudAlignmentService",
     "TrajectoryEvaluationService",
     "compute_trajectory_ape_preview",
     "compute_trajectory_rpe_preview",
@@ -59,7 +58,10 @@ _EVO_ASSOCIATION_MAX_DIFF_S = 0.01
 _SIM3_CLOUD_MIN_MATCHED_PAIRS = 20
 _SIM3_CLOUD_MAX_RMS_ERROR_M = 2.0
 _SIM3_CLOUD_MAX_UP_AXIS_TILT_DEG = 15.0
-_RDF_DOWN_AXIS = np.array([0.0, 1.0, 0.0], dtype=np.float64)
+
+if TYPE_CHECKING:
+    from prml_vslam.pipeline.config import RunConfig
+    from prml_vslam.pipeline.contracts.plan import RunPlan
 
 
 @dataclass(frozen=True, slots=True)
@@ -82,10 +84,6 @@ _POSE_RELATION_UNIT: dict[metrics.PoseRelation, str] = {
     metrics.PoseRelation.translation_part: "m",
     metrics.PoseRelation.rotation_angle_deg: "deg",
 }
-
-if TYPE_CHECKING:
-    from prml_vslam.pipeline.config import RunConfig
-    from prml_vslam.pipeline.contracts.plan import RunPlan
 
 
 @dataclass(slots=True)
@@ -151,10 +149,10 @@ class TrajectoryEvaluationService:
                 f"No matching trajectory timestamps found for Sim(3) alignment (max_diff={_EVO_ASSOCIATION_MAX_DIFF_S:.3f}s)."
             ) from exc
 
-        if not _trajectory_supports_sim3(associated_reference, associated_estimate):
+        if not trajectory_supports_sim3(associated_reference, associated_estimate):
             raise ValueError("Trajectory lacks sufficient geometric spread for Sim(3) alignment.")
 
-        aligned_estimate, alignment = _align_estimate_sim3(
+        aligned_estimate, alignment = align_estimate_sim3(
             reference=associated_reference,
             estimate=associated_estimate,
             max_diff_s=_EVO_ASSOCIATION_MAX_DIFF_S,
@@ -427,56 +425,6 @@ class TrajectoryEvaluationService:
         return run_root / "evaluation" / "trajectory" / "metrics_long.csv"
 
 
-class CloudAlignmentService:
-    """Materialize offline point-cloud alignment artifacts before cloud metrics."""
-
-    def compute_cloud_alignment(self, *, selection: CloudAlignmentSelection) -> CloudAlignmentArtifact:
-        """Refine a trajectory-Sim(3)-aligned cloud against a reference cloud with ICP."""
-        reference_pcd = _read_non_empty_point_cloud(selection.reference_cloud_path, label="reference")
-        estimate_pcd = _read_non_empty_point_cloud(selection.sim3_cloud_path, label="estimate")
-        registration = o3d.pipelines.registration.registration_icp(
-            estimate_pcd,
-            reference_pcd,
-            selection.max_correspondence_distance_m,
-            np.eye(4, dtype=np.float64),
-            o3d.pipelines.registration.TransformationEstimationPointToPoint(),
-        )
-        transformation = np.asarray(registration.transformation, dtype=np.float64)
-        points_xyz, colors_rgb = load_point_cloud_ply_with_colors(selection.sim3_cloud_path)
-        rotation = transformation[:3, :3]
-        translation = transformation[:3, 3]
-        refined_points = points_xyz @ rotation.T + translation
-        icp_cloud_path = self.icp_point_cloud_path(selection.artifact_root)
-        write_point_cloud_ply(icp_cloud_path, refined_points, colors_rgb=colors_rgb)
-        artifact = CloudAlignmentArtifact(
-            path=self.result_path(selection.artifact_root),
-            reference_cloud_path=selection.reference_cloud_path,
-            sim3_point_cloud_path=selection.sim3_cloud_path,
-            icp_point_cloud_path=icp_cloud_path,
-            target_frame=selection.target_frame,
-            max_correspondence_distance_m=selection.max_correspondence_distance_m,
-            fitness=float(registration.fitness),
-            inlier_rmse_m=float(registration.inlier_rmse),
-            transformation=transformation.tolist(),
-        )
-        artifact.path.parent.mkdir(parents=True, exist_ok=True)
-        artifact.path.write_text(
-            json.dumps(artifact.model_dump(mode="json"), indent=2, sort_keys=True),
-            encoding="utf-8",
-        )
-        return artifact
-
-    @staticmethod
-    def result_path(run_root: Path) -> Path:
-        """Return the deterministic point-cloud alignment metadata path."""
-        return run_root / "evaluation" / "cloud_alignment.json"
-
-    @staticmethod
-    def icp_point_cloud_path(run_root: Path) -> Path:
-        """Return the deterministic ICP-refined point-cloud path."""
-        return run_root / "evaluation" / "point_cloud_sim3_icp_aligned.ply"
-
-
 def compute_trajectory_ape_preview(
     *,
     reference_path: Path,
@@ -512,8 +460,8 @@ def compute_trajectory_ape_preview(
     evaluation_estimate = associated_estimate
     alignment = None
     if alignment_mode is TrajectoryAlignmentMode.SIM3_UMEYAMA:
-        if _trajectory_supports_sim3(associated_reference, associated_estimate):
-            evaluation_estimate, alignment = _align_estimate_sim3(
+        if trajectory_supports_sim3(associated_reference, associated_estimate):
+            evaluation_estimate, alignment = align_estimate_sim3(
                 reference=associated_reference,
                 estimate=associated_estimate,
                 max_diff_s=max_diff_s,
@@ -588,54 +536,6 @@ def compute_trajectory_rpe_preview(
     )
 
 
-def _align_estimate_sim3(
-    *,
-    reference: PoseTrajectory3D,
-    estimate: PoseTrajectory3D,
-    max_diff_s: float,
-    target_frame: str = "world",
-    source_frame: str = "slam_world",
-    reference_source: str = "reference",
-    method_id: str | None = None,
-    method_label: str | None = None,
-) -> tuple[PoseTrajectory3D, TrajectoryAlignmentArtifact]:
-    if _is_gravity_aligned_target(target_frame):
-        # Gravity-aligned benchmark worlds (e.g. ADVIO) are near-planar; full
-        # Umeyama can return an up/down-flipped rotation. Lock rotation to yaw
-        # about the RDF gravity axis so the cloud overlay cannot flip upside down.
-        scale_value, rotation_matrix, translation_vector = yaw_similarity_align(
-            np.asarray(estimate.positions_xyz, dtype=np.float64),
-            np.asarray(reference.positions_xyz, dtype=np.float64),
-            up_axis=_RDF_DOWN_AXIS,
-            correct_scale=True,
-        )
-        aligned_estimate = apply_similarity_to_trajectory(
-            estimate, scale=scale_value, rotation=rotation_matrix, translation=translation_vector
-        )
-        rotation, translation, scale = rotation_matrix, translation_vector, scale_value
-    else:
-        aligned_estimate = copy.deepcopy(estimate)
-        rotation, translation, scale = aligned_estimate.align(reference, correct_scale=True)
-    residual = np.asarray(reference.positions_xyz, dtype=np.float64) - np.asarray(
-        aligned_estimate.positions_xyz,
-        dtype=np.float64,
-    )
-    rms_error_m = float(np.sqrt(np.mean(np.sum(residual**2, axis=1))))
-    return aligned_estimate, TrajectoryAlignmentArtifact(
-        source_frame=source_frame,
-        target_frame=target_frame,
-        scale=float(scale),
-        rotation=np.asarray(rotation, dtype=np.float64).tolist(),
-        translation=np.asarray(translation, dtype=np.float64).reshape(3).tolist(),
-        matched_pairs=int(len(reference.positions_xyz)),
-        rms_error_m=rms_error_m,
-        reference_source=reference_source,
-        sync_max_diff_s=max_diff_s,
-        method_id=method_id,
-        method_label=method_label,
-    )
-
-
 def _method_world_frame(method_id: str | None) -> str:
     token = "slam" if method_id is None else _entity_token(str(method_id))
     return f"{token}_slam_world"
@@ -692,7 +592,7 @@ def _apply_sim3_cloud_use_policy(
     *,
     cloud_input_present: bool,
 ) -> TrajectoryAlignmentArtifact:
-    up_axis_tilt_deg = _sim3_up_axis_tilt_deg(np.asarray(alignment.rotation, dtype=np.float64))
+    up_axis_tilt_deg = sim3_up_axis_tilt_deg(np.asarray(alignment.rotation, dtype=np.float64))
     reasons: list[str] = []
     status = TrajectoryAlignmentCloudUseStatus.NOT_REQUESTED
     if cloud_input_present:
@@ -725,37 +625,6 @@ def _apply_sim3_cloud_use_policy(
     )
 
 
-def _sim3_up_axis_tilt_deg(rotation: np.ndarray) -> float | None:
-    if rotation.shape != (3, 3) or not np.all(np.isfinite(rotation)):
-        return None
-    transformed_down_axis = rotation @ _RDF_DOWN_AXIS
-    norm = float(np.linalg.norm(transformed_down_axis))
-    if norm <= 0.0 or not math.isfinite(norm):
-        return None
-    cos_angle = float(np.clip(np.dot(transformed_down_axis / norm, _RDF_DOWN_AXIS), -1.0, 1.0))
-    return math.degrees(math.acos(cos_angle))
-
-
-def _trajectory_supports_sim3(reference: PoseTrajectory3D, estimate: PoseTrajectory3D) -> bool:
-    if len(reference.positions_xyz) < 3 or len(estimate.positions_xyz) < 3:
-        return False
-    reference_centered = np.asarray(reference.positions_xyz, dtype=np.float64) - np.mean(
-        reference.positions_xyz,
-        axis=0,
-    )
-    estimate_centered = np.asarray(estimate.positions_xyz, dtype=np.float64) - np.mean(estimate.positions_xyz, axis=0)
-    return np.linalg.matrix_rank(reference_centered) >= 2 and np.linalg.matrix_rank(estimate_centered) >= 2
-
-
-def _is_gravity_aligned_target(target_frame: str) -> bool:
-    """Whether the benchmark target frame is gravity-aligned (up == RDF -Y).
-
-    ADVIO provider worlds derive from Apple Y-up, so RDF ``-Y`` is gravity. The
-    TUM first-camera RDF frame is *not* gravity-aligned, so it keeps full Umeyama.
-    """
-    return target_frame.startswith("advio_") and target_frame.endswith("_world")
-
-
 def _infer_target_frame(dataset: DatasetId | None, reference_path: Path | None) -> str:
     """Inferred target frame name for UI selections without benchmark input metadata."""
     if dataset is DatasetId.ADVIO:
@@ -783,17 +652,3 @@ def _write_aligned_point_cloud(
     translation = np.asarray(alignment.translation, dtype=np.float64)
     aligned_points = alignment.scale * (points_xyz @ rotation.T) + translation
     write_point_cloud_ply(output_path, aligned_points, colors_rgb=colors_rgb)
-
-
-def _read_non_empty_point_cloud(path: Path, *, label: str) -> Any:
-    if not path.exists():
-        raise FileNotFoundError(f"Point-cloud alignment {label} cloud does not exist: {path}")
-    point_cloud = o3d.io.read_point_cloud(path.as_posix())
-    points_xyz = np.asarray(point_cloud.points, dtype=np.float64)
-    if points_xyz.shape[0] == 0:
-        raise ValueError(f"Point-cloud alignment {label} cloud is empty: {path}")
-    if points_xyz.ndim != 2 or points_xyz.shape[1] != 3:
-        raise ValueError(f"Expected {label} point cloud shape (N, 3), got {points_xyz.shape} for '{path}'.")
-    if not np.isfinite(points_xyz).all():
-        raise ValueError(f"Point-cloud alignment {label} cloud contains non-finite points: {path}")
-    return point_cloud
